Collision detection using character resized hitbox

+++ src/main/java/Utilities/HelpMethods.java
@@ -8,7 +8,7 @@ import com.mycompany.platformgame.Game;
  */
 public class HelpMethods {
     
-    public static boolean CanMoveHere(float x, float y, int width, int height, int[][] lvlData){
+    public static boolean CanMoveHere(float x, float y, float width, float height, int[][] lvlData){
         
         if(!IsSolid(x, y, lvlData))
             if(!IsSolid(x + width, y + height, lvlData))
